Refactor: simplifying table loader examples by using a single TableLoader class
