+++ examples/src/test/scala/com/gu/tableversions/examples/SnapshotTableLoaderSpec.scala
@@ -6,24 +6,38 @@ import java.nio.file.Paths
 import cats.effect.IO
 import com.gu.tableversions.core.TableVersions.{UpdateMessage, UserId}
 import com.gu.tableversions.core._
+import com.gu.tableversions.metastore.Metastore
 import com.gu.tableversions.spark.{SparkHiveMetastore, SparkHiveSuite}
 import org.scalatest.{FlatSpec, Matchers}
 
+/**
+  * This is an example of loading data into a 'snapshot' table, that is, a table where we replace all the content
+  * every time we write to it (no partial updates).
+  */
 class SnapshotTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite {
 
-  import SnapshotTableLoader._
+  import SnapshotTableLoaderSpec._
 
   val table = TableDefinition(TableName(schema, "users"), tableUri, PartitionSchema.snapshot)
+
+  val ddl = s"""CREATE EXTERNAL TABLE IF NOT EXISTS ${table.name.fullyQualifiedName} (
+               |  `id` string,
+               |  `name` string,
+               |  `email` string
+               |)
+               |STORED AS parquet
+               |LOCATION '${table.location}'
+    """.stripMargin
 
   "Writing multiple versions of a snapshot dataset" should "produce distinct versions" in {
     import spark.implicits._
 
-    implicit val tableVersions = InMemoryTableVersions[IO].unsafeRunSync()
-    implicit val metastore = new SparkHiveMetastore[IO]()
+    implicit val tableVersions: TableVersions[IO] = InMemoryTableVersions[IO].unsafeRunSync()
+    implicit val metastore: Metastore[IO] = new SparkHiveMetastore[IO]()
 
     val userId = UserId("test user")
 
-    val loader = new SnapshotTableLoader(table)
+    val loader = new TableLoader[User](table, ddl, isSnapshot = true)
     loader.initTable(userId, UpdateMessage("init"))
 
     // Write the data to the table
@@ -35,7 +49,7 @@ class SnapshotTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite
     loader.insert(identitiesDay1.toDS(), userId, "Committing first version from test")
 
     // Query the table to make sure we have the right data
-    val day1TableData = loader.users().collect()
+    val day1TableData = loader.data().collect()
     day1TableData should contain theSameElementsAs identitiesDay1
 
     // Check underlying storage that the expected versions were written in the right place
@@ -51,7 +65,7 @@ class SnapshotTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite
     loader.insert(identitiesDay2.toDS(), userId, "Committing second version from test")
 
     // Query it to make sure we have the right data
-    val day2TableData = loader.users().collect()
+    val day2TableData = loader.data().collect()
     day2TableData should contain theSameElementsAs identitiesDay2
 
     // Check underlying storage that it was written in the right place
@@ -65,11 +79,11 @@ class SnapshotTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite
 
     // Roll back to previous version
     loader.checkout(versionHistory.drop(1).head.id)
-    loader.users().collect() should contain theSameElementsAs identitiesDay1
+    loader.data().collect() should contain theSameElementsAs identitiesDay1
 
     // Roll forward to latest
     loader.checkout(versionHistory.head.id)
-    loader.users().collect() should contain theSameElementsAs identitiesDay2
+    loader.data().collect() should contain theSameElementsAs identitiesDay2
   }
 
   def versionDirs(tableLocation: URI): List[String] = {
@@ -78,5 +92,11 @@ class SnapshotTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite
     val dir = Paths.get(basePath)
     dir.toFile.list().toList.filter(_.matches(Version.TimestampAndUuidRegex.regex))
   }
+
+}
+
+object SnapshotTableLoaderSpec {
+
+  case class User(id: String, name: String, email: String)
 
 }
